requirements.txt
Flask
yt-dlp
spotipy
python-dotenv
pyinstaller
waitress
pywebview